+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Version 0.23.1-dev
 - Fix type hierarchy calculations when the type no longer is known by the new public API
+- Fix hyphens in packages bug.
 
 ## Version 0.23.0
 - technical: use package_config to interact with package configs